Update headers package and authentication handler***
