+++ httpclient/httpclient_request.go
@@ -164,8 +164,9 @@ func (c *Client) executeRequestWithRetries(method, endpoint string, body, out in
 	}
 
 	// Set request headers
-	log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
-	headerHandler := headers.NewHeaderHandler(req, log, c.APIHandler, c.Token)
+	//log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
+	//headerHandler := headers.NewHeaderHandler(req, log, c.APIHandler, c.Token)
+	headerHandler := headers.NewHeaderHandler(req, c.Logger, c.APIHandler, c.AuthTokenHandler)
 	headerHandler.SetRequestHeaders(endpoint)
 	headerHandler.LogHeaders(c.clientConfig.ClientOptions.HideSensitiveData)
 
@@ -333,8 +334,9 @@ func (c *Client) executeRequest(method, endpoint string, body, out interface{}) 
 	}
 
 	// Set request headers
-	log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
-	headerHandler := headers.NewHeaderHandler(req, log, c.APIHandler, c.Token)
+	//log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
+	//headerHandler := headers.NewHeaderHandler(req, log, c.APIHandler, c.Token)
+	headerHandler := headers.NewHeaderHandler(req, c.Logger, c.APIHandler, c.AuthTokenHandler)
 	headerHandler.SetRequestHeaders(endpoint)
 	headerHandler.LogHeaders(c.clientConfig.ClientOptions.HideSensitiveData)
 
@@ -535,8 +537,8 @@ func (c *Client) DoMultipartRequest(method, endpoint string, fields map[string]s
 	}
 
 	// Initialize HeaderManager
-	log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
-	headerHandler := headers.NewHeaderHandler(req, log, c.APIHandler, c.Token)
+	//log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
+	headerHandler := headers.NewHeaderHandler(req, c.Logger, c.APIHandler, c.AuthTokenHandler)
 
 	// Use HeaderManager to set headers
 	headerHandler.SetContentType(contentType)
